# blog.api/src/main/java/com/benkhalifa/blog/api/model/Contact.java
package com.benkhalifa.blog.api.model;
import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Table;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Entity
@Table(name="contacts")
public class Contact {
	@Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;
	
    @Column
    private String name;
    
    @Column
    private String email;

    @Column
    private String phone;
    
    
    @Column
    private String subject;
    
    @Column
    private String message;




}